Reorganize workspace --> single crate with feature
The main advantage of this switch is that `cargo test` now works again.

Both the pure rust library as well as the python bindings for it are now
implemented in a single crate. This eliminates the need for a cargo
workspace.

By default, the python bindings and related dependencies are not
compiled, which fixes issues when running `cargo test`. To compile the
python bindings, use `cargo build --release --features pybindings`.