Simplify `distributions::Categorical`
Restrict the domain where a `Categorical` distribution is defined to an
interval on `usize` with lower bound zero.

The generalization to domains over different integer types and arbitrary
lower bounds is now done in the python bindings. This is possible now
that the API of the internal rust library consumes and returns only
iterators and no slices, so offsetting the domain on the call site no
longer introduces any overhead over offsetting in the implementation.