Be generic over the precision of probabilities
This uses `#![feature(min_const_generics)]`, which is currently only
available on nightly but expected to be stabilized in Rust 1.50 to be
released on 2021-02-11.

Initial tests suggest that it is indeed helpful to set the `PRECISION`
somewhat lower than the maximum possible value for a given
`CompressedWord` type (see discussion in the documentation of `Coder`).