Revert "Separate encoder state from decoder state"
This reverts the main part of commit
ca0783dacbe5b671106732e2ad7f5f00fe6e79a1.

This is because we won't be implementing `Encode` and `Decode` on the
same type in module `queue` anyway (because interference between the two
once the decoding position comes close to the encoding poisition would
make things unnecessarily complicated). Instead, we'll be providing a
separate decoder "view" into the encoder such that the encoder can't be
used while the decoder is alive.