Fix `Clone` impl of `Stack`
This bug was found by adding test coverage for
`stack::stable::Decoder::into_encoder`, which failed because it clones a
`stable::Decoder`, which invokes `waste.clone` that invalidates the
special invariant of `waste` because it restores the standard `Stack`
invariant for it.

This commit fixes the issue by removing the part that restores the
`Stack` invariant from the `Clone` impl of `Stack`. This means we had to
remove the method `stack::Stable::Coder::waste` because the returned
`&Stack` could now be `clone`d into a potentially invalid `Stack`. This
seems like the cleanest fix since the public API now simply never
exposes any invalid `Stack` (or even a reference to it).