Make `Stack` generic over `CompressedWord`s store
The original `Stack` implementation pushed `CompressedWord`s onto a
`Vec` when encoding and it popped them off the `Vec` when decoding. This
limited the use cases for `Stack` and prevented, e.g., decoding from
immutable slices of `CompressedWord`s as well as implementing `Seek`.

This commit adds a module `stack::backend` that defines traits for
common properties of a store of `CompressedWord`s (e.g., whether or not
reading or writing is possible). These traits are used as bounds for a
generic implementation of `Stack` that can use other storage backends
than `Vec<CompressedWord>`.

This will allow us to implement `Seek` for `Stack` if the backend
storage provides random access.