Fix use of `PhantomData` throughout crate
So far, we only use `PhantomData` to keep track of different invariants
behaviour based on generic arguments. We never actually own any of the
types wrapped in `PhantomData`. Following the documentation of
`PhantomData`, this commit  therefore changes all usages of
`PhantomData<T>` to `PhantomData<*mut T>` (we use `*mut T` rather than
`*const T` to make the resulting types *invariant* rather than covariant
over `T`, which is the more conservative choice and which seems to be
necessary for "marker" generic arguments that enforce invariants, like the
parameter `CompressedWord` of Stack`, which is used to enforce an
invariant on `Stack::state`).