Rename variables with `distribution` to `model`
We now aim to consistently refer to discrete probability distributions
as "models" when they are used for entropy coding. We still use the word
"distribution" for (i) continuous probability distributions and
(ii) when talking about the interours of an entropy model, which
typically requires thinking of the model as a probability distribution.

These changes may contain some false positives and false negatives. But
the renaming has no effect on program behavior, so we'll fix up any
inconsistencies with the new naming convention as we encounter them.