Split `EntropyModel` in `{Encoder, Decoder}Model`
This is useful so that we can implement optimized models that can only do
one of encoding and decoding. For example, for small `PRECISION`, one may
want to implement the quantile function (for decoding) as a lookup
table. But we wouldn't want to support encoding with this representation
of the entropy model because inverting the lookup table for each encoded
symbol would be inefficient.

We still keep a slimmed-down version of `EncoderModel`, on which both
`EncoderModel` and `DecoderModel` depend to ensure that encoding and
decoding use the same `Probability` and `Symbol` types.