Working implementation of Huffman {en, de}coding
This implementation can encode a sequence of symbols to a bitstring
using either a common Huffman tree for all symbols or an individual
Huffman tree for each symbol. The bitstring can then be decoded back
into the same sequence of symbols when using the same sequence of
Huffman trees.

Contains unit tests.