Make range coder lazy in case of inversion
This change means that the range coder no longer needs to be able to
access and modify previously written words. This was the last thing that
kept us from generecising the encoder over its backend.

This change makes the range coder about 3 to 5 % slower in our
benchmarks with the highly optimized lookup models. However, it can
probably be optimized further since this first implemnetation of
lazyness was deliberately kept simple in order to get it right first,
before getting it fast.

This commit also changes how the range encoder seals its bit string when
compression is done (it fills up with zeros rather than ones now). This
doesn't change compression performance but it does change the output of
the encoder in the last few words.