Fix performance regression 1: unsafe nonzero hint
It turns out that rust is not smart enought to compile away the check
for zero in `a / b.get()` where `b` is a `NonZeroU32` or similar.
Introducing the following trivially save hint seems to solve the
problem:

```
let probability = probability.get();
unsafe {
    // SAFETY: `probability` came from a `NonZero`
    if probability == num::zero::<D::Probability>() {
        std::hint::unreachable_unchecked();
    }
}
```

I'll file a bug against the rust repository for this if it's not already
a known bug.