Split `Coder::maybe_empty` into 2
Removes the trait method `Coder::maybe_empty` and adds two new trait
methods:
- `Encoder::maybe_full`; and
- `Decoder::maybe_exhausted`.

Motivation: the trait method `Coder::maybe_empty` made sense for the
(originally only available) implementation of a stack encoder with a
growable/shrinkable backend. However, it is confusing for any other
coders, such as a queue or an `AnsCoder` that is backed by a `Cursor`. What
does "empty" mean for, e.g., a decoder that's backed by a cursor? Is the
whole underlying buffer empty or is there just no compressed data left?

Moving `maybe_empty` from `Code` to `Decode` and renaming it to
`maybe_exhausted` clarifies this. For consistency, this commit also adds
a trait method `maybe_full` to the `Encode` trait. This is useful for
the `stable` module and reflects the API in the `backends` module.