Unify `{Pos, Seek}` from `backends` and `stream`
Removes the `Pos` and `Seek` traits from the `backends` and `stream`
modules and replaces them by a single pair of `Pos` and `Seek` traits at
the top level of the module hierarchy.

Also genericises over the `Position` type (which used to be hardcoded to
`usize`) by introducing a new "parent" trait `PosSeek`. This makes it
possible to treat seeking in backends in the same way as seeking in
coders. It will also be necessary when implementing seeking in more
complex coders (such as the `ChainCoder`) which have more than one
backend and therefore typically need more than one `usize` to specify
the backend position.

TODO: Documentation still needs to be fully updated. While all doc tests
have been ported to work with the new scheme, the text and comments in
some doc tests still refer explicitly to the old `pos_and_state` API.