Renamings in `ChainCoder`
Makes the following renamings for all type parameters and internal
variables in `ChainCoder`:
- `[Qq]uantiles` -> `[Cc]compressed`; and
- `[Rr]emainders` -> `[Rr]emaining`.

This is consistend with the naming of the `from_xxx` and `into_xxx`
methods, and it hides implementation details from the public API (of
which the names of type parameters are a part).

We still use the singular terms `quantile` and `remainder` when
referring to an actual quantile from an entropy model or an actual
remainder from an integer division in the implementations of
`encode_symbol` and `decode_symbol`.