Speed up `LeakilyQuantizedDistribution`
Speeds up the implementation of
`<LeakilyQuantizedDistribution as DecoderModel>::quantile_function`
by replacing the naive linear search with a search that initially
increases the step size exponentially until it finds a second bound on
the symbol so that it can pinpoint the symbol into a finite range. It
then performs a binary search within that range.

All tests still pass and we observe a considerable speedup in extreme
cases, which allowed us to increase `n` in the
`leaky_quantized_binomial` test. But the code is a bit ugly at the
moment, and it may actually slow down more typical cases (still need to
measure).