Another fix for caching in CI
Include rust version in cache key because, apparently, no caches are
saved if the primary key matches. Thus, if we wouldn't include the rust
in the cache key then the cache would become stale as soon as a new Rust
version is out, and it would not get updated until we happen to update
`Cargo.lock`.