Fix compilation errors in no_std mode
It turns out `probability` doesn't work in no_std mode, so we have to
provide mock replacements for the two (tiny) traits on which we rely
from this crate when compiling in no-std mode.

This is not a breaking change because it only changes behavior in no_std
mode, in which, as it turned out, the previous code didn't compile
anyway.

TODO: incorporate test in no-std mode in CI workflow.