Use `maturin-action` Github action for releases
This makes sure best practices are followed for compiling extension
modules. It gives me more confidence in cross-compilation for new Apple
hardware, and it ensures that wheels for linux are compiled in a
"manylinux" container that ensures minimal dependencies on dynamic
libraries.

Once this is merged into master, it should also allow us to trigger a
"dry-run" of the release workflow (through `workflow_dispatch`) that
runs all steps of the release process except for the final steps that
publishing crates, wheels, or documentation.