Fix `pythondoc.py`
This hack is getting even uglier, but it works for now.

For some reason, maturin now creates a python package `constriction`
with the correct submodules (`stream` and `symbol`), but with an
additional submodule `constriction`, which contains the same submodules.
This is due to the generated `__init__.py` file, which imports a module
`constriction` and then imports its submodules recursively.

This new CI script doesn't change the weird submodule setup but it works
around it when generating the Python API reference.