Update python build dependencies
Bump minimum python dependency for building constriction from 3.7
to 3.8, then run `poetry update`. This allows us to build with a newer
numpy release, which fixes a security vulnerability.

This has no effect on the (publicly visible) minimum python or numpy
version for *users* of the constriction python package. It only affects
what is used for *building* (and possibly testing) constriction.