Enable all previously disabled tests on miri
The new version of the `probability` crate no longer uses system calls
for special mathematical functions, so we can now run (almost) all tests
in miri.

The only tests that are still disabled in miri are the benchmarks in
directory `benches`. These cannot be enabled at this point since
`criterion` seems to be incompatible with `miri`.