pybindings: document and check model parameters
Assert that the scale (or equivalent) parameters of a Gaussian, Laplace,
or Cauchy distribution is positive, and state this requirement in the
python API reference. Previously, encoding and decoding with a Laplace
distribution with zero scale did not raise any error but failed to
reproduce the original message in general, see first comment on #38.

Implements the first to-do item in #38.