AugmentedBTree: rename `key` to `pos`
This avoids confusion because both `pos` and `accum` can be used as a
"key" when searching for an entry. Accordingly, the type parameter for
pos is now `P`.

Also, rename `amount` to `count` to indicate that this should probably
be an integer (so that tree restructuring operations don't lead to
rounding errors).