Simplify `AugmentedBTree`
Removes the parent pointers from all nodes in an `AugmentedBTree` as
these turn out to be unnecessary for our use cases. This greatly
simplifies the implementation because it removes aliasing, so we can use
normal owned pointers (`Box`) and no longer need to deal with raw
pointers. It should also make the implementation trivially unwind safe,
which wasn't obviously the case before.