Reimplementation of `BoundedVec`
Implements `BoundedVec<T, CAP>` as a wrapper around
`BoundedPairOfVecs<T, (), CAP>`, and combine the fields `separators` and
`first_child` of `NonLeafNode` into a single field whose type enforces
that the two always have the same length. This makes it a bit easier to
reason about the implementation of `NonLeafNode::insert`.