Unify representation of nodes in `AugmentedBTree`
This will allow us to reduce the amount of code duplication, e.g., for
the implementations of `{NonLeafNode, LeafNode}::remove`.

Use `PairOfBoundedVecs` for both leaf and non-leaf nodes. For non-leaf
nodes, the pair of vecs are the list of separators and their right child
pointers. For leaf nodes, the second bounded vec contains `()` instead
of child pointers. This should completely compile away.

Tests pass, also in miri.