Change allocator dependency in `ensure_no_std`
The allocator provided by the crate `simple-chunk-allocator` lead to
compilation errors for the `ensure_no_std` crate when compiling with
rustc 1.78.0-nightly (f067fd608 2024-02-05). Switching to the allocator
provided by the `talc` crate fixed the issue. It doesn't really matter
much which allocator we use since the `ensure_no_std` crate is just a
dummy crate intended to test whether constriction can be used by a crate
that is compiled in no-std mode (we could probably also use the standard
allocator here, but using an external one seems like a more complete
test of no-std compatibility).