Allow adding/removing multiple points to/from `EmpiricalDistribution`
This has always been supported by the internal tree structure and just
not exposed by the public API because it wasn't clear whether it's
actually useful and we wanted to avoid committing to an implementation
that supports this. But it now turns out that this will be needed by the
python API to implement `EmpiricalDistribution::update_all`, which seems
like evidence that it's a useful feature that's worth exposing publicly.