Allow constructing `EmpiricalDistribution` from `points` and `counts`
Adds this functionality to both the rust and the python API.

See code example in the python documentation of `points_and_counts`.

This also fixes the method `points_and_counts` of the Python API, which
used to return numpy arrays of numpy arrays. It now returns lists of
numpy arrays, which has always been the intention.