Remove Python 3.7 from CI-runs on MacOS
Python 3.7 has been officially unsupported for over half a year now, and
github's CI doesn't seem to support it on MacOS anymore. We're still
going to do Python 3.7 CI-runs on windows and linux for now because they
still seem to work, and the python community is notoriously slow to
retire unsupported python versions.