Implement `_fast` constructors of categ. models
Adds `_fast` and `_perfect` variants for constructors of (precalculated)
categorical entropy models. Both of these construct a model that
approximates given floating point probabilities. The `_perfect` variants
use the old implementation, which explicitly minimizes the
KL-divergence. They have the lowest bit rate but can be very slow. The
`_fast` variants use the (much quicker but very slightly worse)
quantization strategy that the `LazyContiguousCategoricalEntropyModel`
also uses.

The old constructors now delegate to the `_perfect` variants but are
declared deprecated, with disposition to remove in the next breaking
version. The rationale here is that we want to stay binary backward
compatible (thus delegating to the `_perfect` variants), while still
encouraging users to switch to the `_fast` variants whenever backward
compatibility is not necessary.

Still missing:
- most of the documentation;
- extensive unit tests (although this commit already updates all
  existing doc tests to use the new `_fast` variants, and they all
  work).