Temporarily fix `pybindings`
Minimal fixes to make the python bindings for the (new) lazy categorical
distribution work. Mainly meant to make CI pass another step.

Compiling `pybindings` still generates several deprecation warnings
because it currently still uses the legacy constructors to construct
categorical distributions. I'm not fixing this here because the warnings
are a good reminder where more substantive changes to the python API are
necessary to reflect changes in the rust API.