Add `Small` type alias for lookup models
Adds `SmallNonContiguousLookupDecoderModel` and
`SmallContiguousLookupDecoderModel`, and updates code examples in the
documentation to use them (which turns out to simplify them a bit).

This also makes it easier for dependent crates to upgrade to this new
version of `constriction` (for the linguistic flux capacitor, no code
code changes are currently necessary to update to this development
version of `constriction`).