Port PyO3's `text_signature` to `signature`
Part of this was necessary anyway to deal with trailing optional
function arguments in pyo3 0.22, see previous commit. This commit ports
all remaining use-cases of `text_signature` to `signature`, which seems
to be the more modern and better approach, as it type checks the
signatures at compile time, which catches when documentation doesn't
match function signature.

I think the reason I avoided `signature` before were issues with how it
gets rendered in the documentation by `pdoc`, but this seems to be
resolved now (generated documentation looks good).