Fix old documentation
Removes nonsensical documentation from the implementation of a trait
method that was still left over from a very old refactoring
(fortunately, documentations of *implementations of trait methods* are
a bit hidden in the API docs).

This commit also adds analogous documentation to some inherent
methods where it does make sense and should have been to begin with.

Fixes #65.